backend/requirements.txt
fastapi==0.119.0
motor==3.7.1
uvicorn[standard]==0.37.0
pydantic==2.12.2
python-dotenv==1.0.1
